Create Changer.py
Old Changer.py file that would run in command terminal fine but had no real control or ability to monitor.
this file used a Waveshare Epaper display in order to give some monitor capability. waveshare files are not included as not part of the rebuild project yet but will be put back at a later date.